Add support for snapshots CreateFromURL UEFI field (#390)
NOTE: This field was historically a string which took `""`, `"no"`, or `"yes"`. The API has recently
been updated to take `true` and `false` booleans.

This adds the field as a `*bool` so that we can differentiate unset vs explicitly false.

Based on the current unit test setup, there isn't much point in adding coverage for this field.